+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 3.12.2018
+
+* [PlayerActivity] Seek dialog layout setup
+
 #### 2.12.2018
 
 * [PlayerActivity] Basic dialog for video seek